Move humanoids.py to creatures package.
