+++ src/creatures/humanoids.py
