+++ src/dungeon_controller.py
@@ -7,13 +7,13 @@ from pygame.sprite import spritecollide, groupcollide
 
 import abilities
 import controller
-import humanoids
 import mods
 import settings
 import sounds
 import tilemap
 import view
-from humanoids import Player, Mob, collide_hit_rect_with_rect
+from creatures import humanoids
+from creatures.humanoids import Player, Mob, collide_hit_rect_with_rect
 from items.item_manager import ItemManager
 from model import Obstacle, Groups, GameObject, Timer, \
     DynamicObject, Waypoint, Group, ConflictGroups
